feat: 유저정보를 DB에 추가하는 로직을 authResolver에 적용
아울러 inMemoryTokenDB는 불필요해져서 제거하고, 필요한 기능만 남긴 Auth 클래스를 생성함
